transforms understood
Alright. Now I know that transforms are defined as a variable in the dataset definition so I can alter them later. The code I was using to build the dataloader is working with text data which doesn't require transforms. But my image data does and the method the PyTorch team uses to apply the transform and assemble indices is incompatible with this data structure. I now need to resolve this incompatibility by either rewriting my transform application code, or rewriting the code that assembles and appends observation-level information.