added annotation .csv generation
Added an annotation .csv generator so I can avoid attacking the problem from the wrong angle. I now need to debug/finish it, integrate it with the Dataset definition, and figure out how to get this correctly ordered with the Dataloaders. Are there multiple ways of solving this same problem?