consolidating folder references
Consolidated some folder references and adjusted the slashes to conform to Windows settings for the os package. Bug has now moved around. New bug location is in the file renaming loop which is now freshly broken, it appears to be adding the file names onto the previous pathing once it iterates over the first parent folder.