finally identified the error
Traceback was very much unhelpful in identifying the line that was giving the error. Although I found and corrected a number of other things that would later become problems along the way, the problem turned out to essentially be that the randint() method was returning a tensor which couldn't be used in place of an integer to call on an index of a dataframe. Go figure.