reverted to previous version, reference doc
Reverted code to a previous version prior to some troubleshooting messing with my naming. Added a model version of the file based on which I think I can clean this up once and for all. To-do in the future includes chopping the existing document up and breaking it into multiple files.